+++ lib_acl/src/net/listen/acl_unix_listen.c
@@ -34,27 +34,41 @@ ACL_SOCKET acl_unix_listen(const char *addr, int backlog, unsigned flag)
 {
 #undef sun
 	struct sockaddr_un sun;
-	int    len = (int) strlen(addr);
+	size_t len, size;
+	char  *path = sun.sun_path;
 	ACL_SOCKET sock;
-
-	/*
-	 * Translate address information to internal form.
-	 */
-	if (len >= (int) sizeof(sun.sun_path)) {
-		acl_msg_error("%s(%d), %s: unix-domain name too long: %s",
-			__FILE__, __LINE__, __FUNCTION__, addr);
-		return ACL_SOCKET_INVALID;
-	}
 
 	memset((char *) &sun, 0, sizeof(sun));
 	sun.sun_family = AF_UNIX;
+	size = sizeof(sun.sun_path);
+
+#ifdef	ACL_LINUX
+	/* for Linux abstract unix path, we should skip first '@' which was
+	 * marked astract unix in the first of the path by acl.
+	 */
+	if (*addr == '@') {
+		addr++;
+		size--;
+		*path++ = 0;
+	}
+	len = strlen(addr);
+#else
+	len = strlen(addr);
+#endif
+
+	/* Translate address information to internal form. */
+	if (len >= size || len == 0) {
+		acl_msg_error("%s(%d), %s: invalid addr len=%ld, unix path=%s",
+			__FILE__, __LINE__, __FUNCTION__, (long) len, addr);
+		return ACL_SOCKET_INVALID;
+	}
+
 #ifdef HAS_SUN_LEN
 	sun.sun_len = len + 1;
 #endif
-	memcpy(sun.sun_path, addr, len + 1);
+	memcpy(path, addr, len + 1);
 
-	/*
-	 * Create a listener socket. Do whatever we can so we don't run into
+	/* Create a listener socket. Do whatever we can so we don't run into
 	 * trouble when this process is restarted after crash.
 	 */
 	if ((sock = socket(AF_UNIX, SOCK_STREAM, 0)) == ACL_SOCKET_INVALID) {
@@ -63,7 +77,12 @@ ACL_SOCKET acl_unix_listen(const char *addr, int backlog, unsigned flag)
 		return ACL_SOCKET_INVALID;
 	}
 
-	(void) unlink(addr);
+	/* if path is same as sun.sun_path, that's to say the path is not the
+	 * Linux abstract unix
+	 */
+	if (path == sun.sun_path) {
+		(void) unlink(addr);
+	}
 
 	if (bind(sock, (struct sockaddr *) & sun, sizeof(sun)) < 0) {
 		acl_msg_error("%s(%d), %s: bind: %s: %s", __FILE__, __LINE__,
@@ -71,13 +90,15 @@ ACL_SOCKET acl_unix_listen(const char *addr, int backlog, unsigned flag)
 		return ACL_SOCKET_INVALID;
 	}
 #ifdef FCHMOD_UNIX_SOCKETS
-	if (fchmod(sock, 0666) < 0)
+	if (path == sun.sun_path && fchmod(sock, 0666) < 0) {
 		acl_msg_warn("%s(%d), %s: fchmod socket %s: %s", __FILE__,
 			__LINE__, __FUNCTION__, addr, acl_last_serror());
+	}
 #else
-	if (chmod(addr, 0666) < 0)
+	if (path == sun.sun_path && chmod(addr, 0666) < 0) {
 		acl_msg_warn("%s(%d), %s: chmod socket %s: %s", __FILE__,
 			__LINE__, __FUNCTION__, addr, acl_last_serror());
+	}
 #endif
 	acl_non_blocking(sock, flag & ACL_INET_FLAG_NBLOCK ?
 		ACL_NON_BLOCKING : ACL_BLOCKING);
